Restructure folders, create css modules and custom hooks
