+++ frontend/src/pages/user/auth/Signup/Signup.js
@@ -1,49 +1,51 @@
 import React from "react";
 
-import { Link, useNavigate } from "react-router-dom";
+import { Link } from "react-router-dom";
 import { EyeInvisibleOutlined, EyeTwoTone } from "@ant-design/icons";
-import { Row, Col, Button, Form, Input, message } from "antd";
+import { Row, Col, Button, Form, Input } from "antd";
 
-import api from "../helpers/api";
-import loginIllustration from "../images/login_illustration.svg";
+import { useSignup } from "./hooks/useSignup";
+import signupIllustration from "../../../../images/login_illustration.svg";
 
-import AuthLayout from "../layouts/AuthLayout";
+import AuthLayout from "../../../../shared/layouts/AuthLayout";
 
-const Login = () => {
-  const navigate = useNavigate();
-  const login = async (values) => {
-    const { email, password } = values;
-    const { data } = await api.post("/login", { email, password });
+const Signup = () => {
+  const { register } = useSignup();
 
-    if (data.meta.code === 200) {
-      const { data: userData } = data;
-      const { user } = userData;
-
-      localStorage.setItem(
-        "user",
-        JSON.stringify({
-          id: user.id,
-          first_name: user.first_name,
-          last_name: user.last_name,
-          email: user.email,
-          user_type: user.user_type,
-          avatar_url: user.avatar_url,
-        })
-      );
-
-      return navigate("/");
-    }
-
-    message.error(data.meta.message);
-  };
   return (
-    <AuthLayout illustration={loginIllustration}>
+    <AuthLayout illustration={signupIllustration}>
       <div style={{ textAlign: "center", width: "min(25vw, 400px)" }}>
-        <Form onFinish={login} autoComplete="off">
-          <h1>Sign In</h1>
+        <Form onFinish={register} autoComplete="off">
+          <h1>Sign Up</h1>
           <h4 style={{ color: "#8C8C8C", marginBottom: "20px" }}>
             Welcome to E-Learning App
           </h4>
+          <Form.Item
+            name="first_name"
+            rules={[
+              { required: true, message: "Please input your first name" },
+              {
+                whitespace: true,
+                message: "Missing Value",
+              },
+            ]}
+          >
+            <Input className="form-input" placeholder="First Name" />
+          </Form.Item>
+
+          <Form.Item
+            name="last_name"
+            rules={[
+              { required: true, message: "Please input your last name" },
+              {
+                whitespace: true,
+                message: "Missing Value",
+              },
+            ]}
+          >
+            <Input className="form-input" placeholder="Last Name" />
+          </Form.Item>
+
           <Form.Item
             name="email"
             rules={[
@@ -54,11 +56,7 @@ const Login = () => {
               },
             ]}
           >
-            <Input
-              className="form-input"
-              placeholder="Email Address"
-              style={{ padding: "0.7em 1em" }}
-            />
+            <Input className="form-input" placeholder="Email Address" />
           </Form.Item>
 
           <Form.Item
@@ -87,7 +85,7 @@ const Login = () => {
             style={{ borderRadius: "7px", height: "3em" }}
             block
           >
-            Login
+            Register
           </Button>
         </Form>
         <Row style={{ marginTop: "10px" }} gutter={5}>
@@ -98,13 +96,13 @@ const Login = () => {
                 marginBottom: 0,
               }}
             >
-              Don't have an account yet?
+              Already have an account?
             </h4>
           </Col>
           <Col>
-            <Link to="/signup">
+            <Link to="/login" className="center">
               <Button type="link" style={{ padding: 0 }}>
-                Sign Up!
+                Sign In!
               </Button>
             </Link>
           </Col>
@@ -114,4 +112,4 @@ const Login = () => {
   );
 };
 
-export default Login;
+export default Signup;
