+++ frontend/src/pages/user/home/Profile/Profile.module.css
@@ -1,0 +1,37 @@
+.avatar {
+  background-color: #E6F7FF;
+  width: 100%;
+  height: auto;
+  aspect-ratio: 1 / 1;
+}
+
+.userDetails {
+  display: flex;
+  align-items: center;
+  flex-direction: column;
+}
+
+.followRow {
+  margin-top: 0.75em;
+  width: 100%;
+}
+
+.followPoint {
+  cursor: pointer;
+}
+
+.followBtn {
+  margin-top: 1.5em;
+  border-radius: 1.5rem;
+  width: 100%;
+}
+
+.bigText {
+  font-size: 1.5rem;
+}
+
+.learnedLessons {
+  margin: 0.5rem 0;
+  width: 100%;
+  cursor: default;
+}
