+++ frontend/src/shared/components/AuthIllustration.js
